requirements.txt
flask==3.0.0
flask-cors==5.0.0
requests==2.31.0
gunicorn==22.0.0
python-dotenv==1.0.1
